Db and sensor model changes
Add new columns to sensor model and handle them in manager, validator, and DAL
Add new table trouble code
Add new columns to reading table
Fix problems with mapping

+++ IntelligentDiagnostician.BL/Manager/ReadingManager/IReadingManager.cs
@@ -2,5 +2,5 @@
 
 public interface IReadingManager
 {
-    Task CreateAsync(int sensorId, Guid userId, int value);
+    Task CreateAsync(int sensorId, Guid userId, float value);
 }
